+++ src/renderer/src/components/ModelLibrary/providers/index.ts
@@ -1,5 +1,6 @@
 export * from './AnthropicParameters';
 export * from './AnthropicCompatibleParameters';
+export * from './AuggieParameters';
 export * from './AzureParameters';
 export * from './BedrockParameters';
 export * from './CerebrasParameters';